Fix animation not syncing with audio after scrolling back onto the page
If the metronome scrolls off the page, and then back on, the visual CSS
animation may no longer sync up with the audio playing. This fixes that
by introducing a useCSSAnimationSynchronizer hook that the three
components currently using animation (Metronome, MetronomeIcon,
StyleLine) now use to sync up their animations with an ongoing timer.

This could probably be improved in the future by not waiting for a
nextBeat and instead figuring out the exact time until the next beat and
reset the animation to that point. Currently when the metronome
component becomes visible again, the animations will remain frozen
(`animation: none`) until a next beat happens, which is not too
noticeable on fast tempos but fairly noticeable on slower tempos.